requirements.txt
django >=2.0.4
psycopg2 >=2.7.4
pillow >=5.1.0
djangorestframework==3.8.2